pyproject.toml
[tool.poetry]
name = "discord-link-reply"
version = "0.1.0"
description = ""
authors = ["Ellie T <ellie@eltariel.com>"]
readme = "README.md"
packages = [{include = "discord_link_reply"}]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
